[#10] Move worker files into `src` module
